# sql: reuse builtin instances when computing window functions