Previously, windowers would create a new instance of a builtin for
every partition and for every window function which is quite
inefficient in terms of memory usage.

Reset() method is now added to aggregate and window functions
builtins, and now windowers can create a single instance for each
window function and reuse it later on every partition.

Fixes: #34316.

Release note: None