@yuzefovich 2019-01-30
Here are some benchmarks (there is no noticeable difference in performance, but memory usage in the same benchmark is reduced by about 0.36GB with this change).

Before:
```
BenchmarkWindower/SUM()_OVER_()-12                   30      40054064 ns/op      59.92 MB/s
BenchmarkWindower/SUM()_OVER_(ORDER_BY)-12           20      74215417 ns/op      32.34 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12                   10     210519698 ns/op      11.40 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12                    5     207230296 ns/op      11.58 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12                         30      42603904 ns/op      56.33 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12                30      46934311 ns/op      51.14 MB/s
```

After:
```
BenchmarkWindower/SUM()_OVER_()-12                   30      41187395 ns/op      58.27 MB/s
BenchmarkWindower/SUM()_OVER_(ORDER_BY)-12           20      75526073 ns/op      31.78 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/)-12                   10     154642613 ns/op      15.52 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_SINGLE_ROW_PARTITIONS_*/_ORDER_BY)-12                   10     158043560 ns/op      15.19 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/)-12                         30      43865241 ns/op      54.71 MB/s
BenchmarkWindower/SUM()_OVER_(PARTITION_BY/*_MULTIPLE_ROWS_PARTITIONS_*/_ORDER_BY)-12                30      47840714 ns/op      50.17 MB/s
```